# Conflicts:
#	.idea/dataSources/6550f674-7897-41a3-af68-ae36142d58f9.xml
#	README.md
#	library.db
#	src/AudioVideoMaterial.java
#	src/Book.java
#	src/Database.java
#	src/Document.java
#	src/JournalArticle.java
#	src/Main.java

+++ src/AudioVideoMaterial.java
@@ -10,7 +10,7 @@ public class AudioVideoMaterial extends Document{
      * @param Price
      * @param KeyWords
      */
-    public AudioVideoMaterial(int ID, String Title, String Authors, boolean IsAllowedForStudents, int NumberOfCopies, boolean IsReference, float Price, String KeyWords){
+    public AudioVideoMaterial(int ID, String Title, String Authors, boolean IsAllowedForStudents, int NumberOfCopies, boolean IsReference, double Price, String KeyWords){
         super(ID, Title, Authors, IsAllowedForStudents, NumberOfCopies, IsReference, Price, KeyWords);
     }
 }
