+++ src/Book.java
@@ -23,7 +23,7 @@ public class Book extends Document{
      * @param Price
      * @param KeyWords
      */
-    public Book(int ID, String Title, String Authors, boolean IsAllowedForStudents, int NumberOfCopies, boolean IsReference, float Price, String KeyWords, String Publisher, int Edition, boolean IsBestseller){
+    public Book(int ID, String Title, String Authors, boolean IsAllowedForStudents, int NumberOfCopies, boolean IsReference, double Price, String KeyWords, String Publisher, int Edition, boolean IsBestseller){
         super(ID, Title, Authors, IsAllowedForStudents, NumberOfCopies, IsReference, Price, KeyWords);
         setPublisher(Publisher);
         setEdition(Edition);
